Move footer to a component
