+++ frontend/pages/index.js
@@ -36,6 +36,8 @@ import {
 
 import { parse } from "mathjs";
 
+import {Footer} from "../components/footer";
+
 ///////////////////// Add your benchmarks here /////////////////////
 
 // Import our benchmark results
@@ -428,28 +430,7 @@ const Home = () => {
     }
   };
 
-  const Authors = () => {
-    // only two authors for now
-    const _authors = [
-      { name: "George Kadianakis", website: "https://github.com/asn-d6" },
-      { name: "Michele Orrù", website: "https://tumbolandia.net" },
-    ];
-    const [authors, setAuthors] = React.useState(_authors);
 
-    // Randomize author list
-    useEffect(() => {
-      const randomizedAuthors = [..._authors].sort(() => 0.5 - Math.random());
-      setAuthors(randomizedAuthors);
-    }, []);
-
-    return (
-      <Text style={{ fontSize: 12 }}>
-        Developed by
-        <a href={authors[0].website}> {authors[0].name}</a> and{" "}
-        <a href={authors[1].website}>{authors[1].name}</a>.
-      </Text>
-    );
-  };
 
   const validateQuantity = async (rule, value) => {
     if (value.trim() === "") {
@@ -649,9 +630,7 @@ const Home = () => {
             }}
           />
         </Row>
-      <div id="footer">
-        <Authors />
-      </div>
+      <Footer/>
       </Layout.Content>
     </Layout>
   );
